IECoreCycles : Refactor computeTangents to live in AttributeAlgo, but later look into https://developer.blender.org/D15589
