+++ src/GafferCycles/IECoreCyclesPreview/MeshAlgo.cpp
@@ -52,311 +52,11 @@
 #include "util/param.h"
 #include "util/types.h"
 
-// MikkTspace
-#include "Mikktspace/mikktspace.h"
-
 using namespace std;
 using namespace Imath;
 using namespace IECore;
 using namespace IECoreScene;
 using namespace IECoreCycles;
-
-namespace
-{
-
-struct MikkUserData {
-	MikkUserData( const char *layer_name,
-				  ccl::Mesh *mesh,
-				  ccl::float3 *tangent,
-				  float *tangent_sign )
-		: mesh( mesh ), texface( nullptr ), tangent( tangent ), tangent_sign( tangent_sign )
-	{
-		const ccl::AttributeSet &attributes = (mesh->get_num_subd_faces()) ? mesh->subd_attributes :
-																		  mesh->attributes;
-
-		ccl::Attribute *attr_vN = attributes.find( ccl::ATTR_STD_VERTEX_NORMAL );
-#ifdef WITH_CYCLES_CORNER_NORMALS
-		ccl::Attribute* attr_cN = attributes.find( ccl::ATTR_STD_CORNER_NORMAL );
-		if( !attr_vN && !attr_cN )
-#else
-		if( !attr_vN )
-#endif
-		{
-			mesh->add_face_normals();
-			mesh->add_vertex_normals();
-			attr_vN = attributes.find( ccl::ATTR_STD_VERTEX_NORMAL );
-		}
-
-#ifdef WITH_CYCLES_CORNER_NORMALS
-		// This preference depends on what Cycles does inside the hood.
-		// Works for now, but there should be a more clear way of knowing
-		// which normals are used for rendering.
-		if (attr_cN)
-		{
-			corner_normal = attr_cN->data_float3();
-		}
-		else
-		{
-			vertex_normal = attr_vN->data_float3();
-		}
-#else
-		vertex_normal = attr_vN->data_float3();
-#endif
-
-		ccl::Attribute *attr_uv = attributes.find( ccl::ustring( layer_name ) );
-		if( attr_uv != nullptr )
-		{
-			texface = attr_uv->data_float2();
-		}
-	}
-
-	ccl::Mesh *mesh;
-	int num_faces;
-
-	ccl::float3* corner_normal;
-	ccl::float3 *vertex_normal;
-	ccl::float2 *texface;
-	ccl::float3 *orco;
-	ccl::float3 orco_loc, orco_size;
-
-	ccl::float3 *tangent;
-	float *tangent_sign;
-};
-
-int mikk_get_num_faces( const SMikkTSpaceContext *context )
-{
-	const MikkUserData *userdata = (const MikkUserData *)context->m_pUserData;
-	if( userdata->mesh->get_num_subd_faces() )
-	{
-		return userdata->mesh->get_num_subd_faces();
-	}
-	else
-	{
-		return userdata->mesh->num_triangles();
-	}
-}
-
-int mikk_get_num_verts_of_face( const SMikkTSpaceContext *context, const int face_num )
-{
-	const MikkUserData *userdata = (const MikkUserData *)context->m_pUserData;
-	if( userdata->mesh->get_num_subd_faces() )
-	{
-		const ccl::Mesh *mesh = userdata->mesh;
-		return mesh->get_subd_num_corners()[face_num];
-	}
-	else
-	{
-		return 3;
-	}
-}
-
-int mikk_vertex_index( const ccl::Mesh *mesh, const int face_num, const int vert_num )
-{
-	if( mesh->get_num_subd_faces() )
-	{
-		const ccl::Mesh::SubdFace &face = mesh->get_subd_face(face_num);
-		return mesh->get_subd_face_corners()[face.start_corner + vert_num];
-	}
-	else
-	{
-		return mesh->get_triangles()[face_num * 3 + vert_num];
-	}
-}
-
-int mikk_corner_index( const ccl::Mesh *mesh, const int face_num, const int vert_num )
-{
-	if( mesh->get_num_subd_faces() )
-	{
-		const ccl::Mesh::SubdFace &face = mesh->get_subd_face(face_num);
-		return face.start_corner + vert_num;
-	}
-	else
-	{
-		return face_num * 3 + vert_num;
-	}
-}
-
-void mikk_get_position( const SMikkTSpaceContext *context,
-							   float P[3],
-							   const int face_num,
-							   const int vert_num )
-{
-	const MikkUserData *userdata = (const MikkUserData *)context->m_pUserData;
-	const ccl::Mesh *mesh = userdata->mesh;
-	const int vertex_index = mikk_vertex_index(mesh, face_num, vert_num);
-	const ccl::float3 vP = mesh->get_verts()[vertex_index];
-	P[0] = vP.x;
-	P[1] = vP.y;
-	P[2] = vP.z;
-}
-
-void mikk_get_texture_coordinate( const SMikkTSpaceContext *context,
-										 float uv[2],
-										 const int face_num,
-										 const int vert_num )
-{
-	const MikkUserData *userdata = (const MikkUserData *)context->m_pUserData;
-	const ccl::Mesh *mesh = userdata->mesh;
-	if( userdata->texface != nullptr )
-	{
-		const int corner_index = mikk_corner_index( mesh, face_num, vert_num );
-		ccl::float2 tfuv = userdata->texface[corner_index];
-		uv[0] = tfuv.x;
-		uv[1] = tfuv.y;
-	}
-	else if (userdata->orco != nullptr)
-	{
-		const int vertex_index = mikk_vertex_index(mesh, face_num, vert_num);
-		const ccl::float3 orco_loc = userdata->orco_loc;
-		const ccl::float3 orco_size = userdata->orco_size;
-		const ccl::float3 orco = (userdata->orco[vertex_index] + orco_loc) / orco_size;
-
-		const ccl::float2 tmp = map_to_sphere(orco);
-		uv[0] = tmp.x;
-		uv[1] = tmp.y;
-	}
-	else
-	{
-		uv[0] = 0.0f;
-		uv[1] = 0.0f;
-	}
-}
-
-void mikk_get_normal( const SMikkTSpaceContext *context,
-							 float N[3],
-							 const int face_num,
-							 const int vert_num)
-{
-	const MikkUserData *userdata = (const MikkUserData *)context->m_pUserData;
-	const ccl::Mesh *mesh = userdata->mesh;
-	ccl::float3 vN;
-	if( mesh->get_num_subd_faces() )
-	{
-		const ccl::Mesh::SubdFace &face = mesh->get_subd_face(face_num);
-		if (userdata->corner_normal)
-		{
-			vN = userdata->corner_normal[face.start_corner + vert_num];
-		}
-		else if( face.smooth )
-		{
-			const int vertex_index = mikk_vertex_index( mesh, face_num, vert_num );
-			vN = userdata->vertex_normal[vertex_index];
-		}
-		else
-		{
-			vN = face.normal( mesh );
-		}
-	}
-	else
-	{
-		if (userdata->corner_normal)
-		{
-			vN = userdata->corner_normal[face_num * 3 + vert_num];
-		}
-		if( mesh->get_smooth()[face_num] )
-		{
-			const int vertex_index = mikk_vertex_index( mesh, face_num, vert_num );
-			vN = userdata->vertex_normal[vertex_index];
-		}
-		else
-		{
-			const ccl::Mesh::Triangle tri = mesh->get_triangle( face_num );
-			vN = tri.compute_normal(&mesh->get_verts()[0]);
-		}
-	}
-	N[0] = vN.x;
-	N[1] = vN.y;
-	N[2] = vN.z;
-}
-
-void mikk_set_tangent_space(const SMikkTSpaceContext *context,
-								   const float T[],
-								   const float sign,
-								   const int face_num,
-								   const int vert_num)
-{
-	MikkUserData *userdata = (MikkUserData *)context->m_pUserData;
-	const ccl::Mesh *mesh = userdata->mesh;
-	const int corner_index = mikk_corner_index( mesh, face_num, vert_num );
-	userdata->tangent[corner_index] = ccl::make_float3( T[0], T[1], T[2] );
-	if (userdata->tangent_sign != nullptr)
-	{
-		userdata->tangent_sign[corner_index] = sign;
-	}
-}
-
-void mikk_compute_tangents( const char *layer_name, ccl::Mesh *mesh, bool need_sign, bool active_render )
-{
-	/* Create tangent attributes. */
-	ccl::AttributeSet &attributes = ( mesh->get_num_subd_faces() ) ? mesh->subd_attributes : mesh->attributes;
-	ccl::Attribute *attr;
-	ccl::ustring name;
-
-	if (layer_name != nullptr)
-	{
-		name = ccl::ustring( ( std::string( layer_name ) + ".tangent" ).c_str() );
-	}
-	else
-	{
-		name = ccl::ustring( "orco.tangent" );
-	}
-
-	if( active_render )
-	{
-		attr = attributes.add( ccl::ATTR_STD_UV_TANGENT, name);
-	}
-	else
-	{
-		attr = attributes.add( name, ccl::TypeDesc::TypeVector, ccl::ATTR_ELEMENT_CORNER );
-	}
-	ccl::float3 *tangent = attr->data_float3();
-	/* Create bitangent sign attribute. */
-	float *tangent_sign = nullptr;
-	if( need_sign )
-	{
-		ccl::Attribute *attr_sign;
-		ccl::ustring name_sign;
-
-		if (layer_name != nullptr)
-		{
-			name_sign = ccl::ustring( ( std::string( layer_name ) + ".tangent_sign" ).c_str() );
-		}
-		else
-		{
-			name_sign = ccl::ustring( "orco.tangent_sign" );
-		}
-
-		if( active_render )
-		{
-			attr_sign = attributes.add( ccl::ATTR_STD_UV_TANGENT_SIGN, name_sign );
-		}
-		else
-		{
-			attr_sign = attributes.add( name_sign, ccl::TypeDesc::TypeFloat, ccl::ATTR_ELEMENT_CORNER );
-		}
-		tangent_sign = attr_sign->data_float();
-	}
-	/* Setup userdata. */
-	MikkUserData userdata( layer_name, mesh, tangent, tangent_sign );
-	/* Setup interface. */
-	SMikkTSpaceInterface sm_interface;
-	memset( &sm_interface, 0, sizeof( sm_interface ) );
-	sm_interface.m_getNumFaces = mikk_get_num_faces;
-	sm_interface.m_getNumVerticesOfFace = mikk_get_num_verts_of_face;
-	sm_interface.m_getPosition = mikk_get_position;
-	sm_interface.m_getTexCoord = mikk_get_texture_coordinate;
-	sm_interface.m_getNormal = mikk_get_normal;
-	sm_interface.m_setTSpaceBasic = mikk_set_tangent_space;
-	/* Setup context. */
-	SMikkTSpaceContext context;
-	memset( &context, 0, sizeof( context ) );
-	context.m_pUserData = &userdata;
-	context.m_pInterface = &sm_interface;
-	/* Compute tangents. */
-	genTangSpaceDefault( &context );
-}
-
-} // namespace
 
 namespace
 {
@@ -1102,19 +802,6 @@ ccl::Object *convert( const std::vector<const IECoreScene::MeshPrimitive *> &mes
 	cobject->set_geometry( static_cast<ccl::Geometry*>( cmesh ) );
 	cobject->name = ccl::ustring(nodeName.c_str());
 	return cobject;
-}
-
-void computeTangents( ccl::Mesh *cmesh, const IECoreScene::MeshPrimitive *mesh, bool needsign )
-{
-	const ccl::AttributeSet &attributes = (cmesh->get_num_subd_faces()) ? cmesh->subd_attributes :
-																		  cmesh->attributes;
-
-	ccl::Attribute *attr = attributes.find( ccl::ATTR_STD_UV );
-	if( attr )
-		mikk_compute_tangents( attr->standard_name( ccl::ATTR_STD_UV ), cmesh, needsign, true );
-
-	// Secondary UVsets
-	// TODO: Currently seeing Cycles not working with normal maps and secondary UVs.
 }
 
 } // namespace MeshAlgo
